# dockstore-webservice/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: This describes the dockstore API, a webservice that manages pairs of Docker
    images and associated metadata such as CWL documents and Dockerfiles used to
    build those images
  version: 1.7.0-alpha.2-SNAPSHOT
  title: Dockstore API
  termsOfService: TBD
  contact:
    name: Dockstore@ga4gh
    url: https://github.com/ga4gh/dockstore
    email: theglobalalliance@genomicsandhealth.org
  license:
    name: Apache License Version 2.0
    url: https://github.com/ga4gh/dockstore/blob/develop/LICENSE
...
paths:
...
  /api/ga4gh/v2/extended:
    get:
      tags:
        - extendedGA4GH
      summary: List all published services, optionally filtering by subclass
      description: ""
      operationId: getPublishedServices
      parameters:
        - name: subclass
          in: query
          description: Filter by subclass
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
...
  "/api/ga4gh/v2/extended/{serviceId}":
    get:
      tags:
        - extendedGA4GH
      summary: Fetches the specified service version
      description: ""
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
            text/plain:
              schema:
                $ref: "#/components/schemas/Service"
  "/api/ga4gh/v2/extended/{serviceId}/versions":
    get:
      tags:
        - extendedGA4GH
      summary: Gets all versions of the specified version
      description: ""
      operationId: getServiceVersions
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
  "/api/ga4gh/v2/extended/{serviceId}/versions/{versionId}":
    get:
      tags:
        - extendedGA4GH
      summary: Gets the specified service version
      description: ""
      operationId: getServiceVersion
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceVersion"
            text/plain:
              schema:
                $ref: "#/components/schemas/ServiceVersion"
...
  /services:
    get:
      tags:
        - services
      summary: List all published services, optionally filtering by subclass
      description: ""
      operationId: getPublishedServices
      parameters:
        - name: subclass
          in: query
          description: Filter by subclass
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  "/services/{serviceId}":
    get:
      tags:
        - services
      summary: Fetches the specified service version
      description: ""
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  "/services/{serviceId}/versions":
    get:
      tags:
        - services
      summary: Gets all versions of the specified version
      description: ""
      operationId: getServiceVersions
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
  "/services/{serviceId}/versions/{versionId}":
    get:
      tags:
        - services
      summary: Gets the specified service version
      description: ""
      operationId: getServiceVersion
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceVersion"
...
servers:
  - url: https://dockstore.org/api
components:
...
  securitySchemes:
    BEARER:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Implementation specific ID for the service in this web service
        subClass:
          type: string
          description: The type of service
          enum:
            - DOCKER_COMPOSE
            - SWARM
            - KUBERNETES
            - HELM
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ServiceVersion"
        gitUrl:
          type: string
          description: The git URL of this service
    ServiceVersion:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Implementation specific ID for the service version in this web
            service
        subClass:
          type: string
          description: The subclass of this service
          enum:
            - DOCKER_COMPOSE
            - SWARM
            - KUBERNETES
            - HELM
        sourceFiles:
          type: array
          description: The .dockstore.yml, test parameter file(s), and readme for this
            service.
          uniqueItems: true
          items:
            $ref: "#/components/schemas/SourceFile"
        commitId:
          type: string
          description: The service version's Git commit id
...
    SourceFile:
      type: object
      required:
        - absolutePath
        - path
        - type
      properties:
        id:
          type: integer
          format: int64
          description: Implementation specific ID for the source file in this web service
        verifiedBySource:
          type: object
          description: maps from platform to whether an entry successfully ran on it using
            this test json
          additionalProperties:
            $ref: "#/components/schemas/VerificationInformation"
        type:
          type: string
          description: Enumerates the type of file
          enum:
            - DOCKSTORE_CWL
            - DOCKSTORE_WDL
            - DOCKERFILE
            - CWL_TEST_JSON
            - WDL_TEST_JSON
            - NEXTFLOW
            - NEXTFLOW_CONFIG
            - NEXTFLOW_TEST_PARAMS
            - DOCKSTORE_YML
        content:
          type: string
          description: Cache for the contents of the target file
        path:
          type: string
          description: Path to sourcefile relative to its parent
        absolutePath:
          type: string
          description: Absolute path of sourcefile in git repo
...
    VerificationInformation:
      type: object
      properties:
        verified:
          type: boolean
        metadata:
          type: string
        platformVersion:
          type: string
...